# backend/felcn-base-backend/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Backend Base NestJS
  version: '1.0'
  description: Documentación Backend base - NestJS
  contact:
    name: agetic@agetic.gob.bo
servers:
  - url: 'https://base.agcs.agetic.gob.bo/ws/api'
paths:
...
  /usuarios:
    get:
      operationId: UsuarioController_listar
      summary: API para obtener el listado de usuarios
      parameters:
        - name: limite
          required: false
          in: query
          schema:
            minimum: 10
            maximum: 50
            default: 10
            type: number
        - name: pagina
          required: false
          in: query
          schema:
            minimum: 1
            default: 1
            type: number
        - name: filtro
          required: false
          in: query
          schema:
            type: string
        - name: ordenRaw
          required: false
          in: query
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: &ref_8
        - Usuarios
      security:
        - bearer: []
    post:
      operationId: UsuarioController_crear
      summary: API para crear un nuevo usuario
      parameters: []
      requestBody:
        required: true
        description: new Usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUsuarioDto'
      responses:
        '201':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/cuenta/perfil:
    get:
      operationId: UsuarioController_obtenerPerfil
      summary: Obtiene la información del perfil del usuario
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/crear-cuenta:
    post:
      operationId: UsuarioController_crearUsuario
      summary: API para crear una nueva Cuenta
      parameters: []
      requestBody:
        required: true
        description: Nueva cuenta de usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUsuarioDto'
      responses:
        '201':
          description: ''
      tags: *ref_8
  /usuarios/recuperar:
    post:
      operationId: UsuarioController_recuperarCuenta
      summary: API para recuperar una Cuenta
      parameters: []
      requestBody:
        required: true
        description: Cuenta
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecuperarCuentaDto'
      responses:
        '201':
          description: ''
      tags: *ref_8
  /usuarios/validar-recuperar:
    post:
      operationId: UsuarioController_validarRecuperarCuenta
      summary: API para validar recuperación una Cuenta
      parameters: []
      requestBody:
        required: true
        description: Cuenta
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidarRecuperarCuentaDto'
      responses:
        '201':
          description: ''
      tags: *ref_8
  /usuarios/cuenta/activacion:
    patch:
      operationId: UsuarioController_activarCuenta
      summary: API para activar una Cuenta
      parameters: []
      requestBody:
        required: true
        description: Cuenta
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivarCuentaDto'
      responses:
        '200':
          description: ''
      tags: *ref_8
  /usuarios/cuenta/nueva-contrasena:
    patch:
      operationId: UsuarioController_nuevaContrasena
      summary: API para nueva Contraseña
      parameters: []
      requestBody:
        required: true
        description: Cuenta
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuevaContrasenaDto'
      responses:
        '200':
          description: ''
      tags: *ref_8
  /usuarios/cuenta/ciudadania:
    post:
      operationId: UsuarioController_crearConCiudadania
      summary: API para crear un nuevo usuario relacionado con Ciudadanía Digital
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUsuarioCiudadaniaDto'
      responses:
        '201':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/{id}/activacion:
    patch:
      operationId: UsuarioController_activar
      summary: Activa un usuario
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/{id}/inactivacion:
    patch:
      operationId: UsuarioController_inactivar
      summary: Inactiva un usuario
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/cuenta/contrasena:
    patch:
      operationId: UsuarioController_actualizarContrasena
      summary: Actualiza la contrasena de un usuario authenticado
      parameters: []
      requestBody:
        required: true
        description: new Rol
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarContrasenaDto'
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/{id}/restauracion:
    patch:
      operationId: UsuarioController_restaurarContrasena
      summary: API para restaurar la contraseña de un usuario
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/{id}/reenviar:
    patch:
      operationId: UsuarioController_reenviarCorreoActivacion
      summary: API para reenviar Correo Activación
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
  /usuarios/{id}:
    patch:
      operationId: UsuarioController_actualizarDatos
      summary: Actualiza datos de un usuario
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      requestBody:
        required: true
        description: Usuario
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarUsuarioRolDto'
      responses:
        '200':
          description: ''
      tags: *ref_8
      security:
        - bearer: []
...
  /usuarios/test/codigo/{id}:
    get:
      operationId: UsuarioController_obtenerCodigo
      parameters:
        - name: id
          required: true
          in: path
          example: '12'
          schema:
            type: string
      responses:
        '200':
          description: ''
      tags: *ref_8
...
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PersonaDto:
      type: object
      properties:
        nroDocumento:
          type: string
          example: '4192299'
        nombres:
          type: string
          example: MARIELA
        primerApellido:
          type: string
          example: ALCAZAR
        segundoApellido:
          type: string
          example: ALMARAZ
        fechaNacimiento:
          type: object
          example: '2002-05-04'
      required:
        - nroDocumento
        - nombres
        - primerApellido
        - segundoApellido
        - fechaNacimiento
    CrearUsuarioDto:
      type: object
      properties:
        correoElectronico:
          type: string
          example: 123456@gmail.com
        persona:
          $ref: '#/components/schemas/PersonaDto'
        roles:
          example:
            - '1'
          type: array
          items:
            type: string
      required:
        - correoElectronico
        - persona
        - roles
    RecuperarCuentaDto:
      type: object
      properties:
        correoElectronico:
          type: string
          example: 123456@gmail.com
      required:
        - correoElectronico
    ValidarRecuperarCuentaDto:
      type: object
      properties:
        codigo:
          type: string
      required:
        - codigo
    ActivarCuentaDto:
      type: object
      properties:
        codigo:
          type: string
      required:
        - codigo
    NuevaContrasenaDto:
      type: object
      properties:
        codigo:
          type: string
        contrasenaNueva:
          type: string
      required:
        - codigo
        - contrasenaNueva
    CrearUsuarioCiudadaniaDto:
      type: object
      properties:
        usuario:
          type: string
          example: '123456'
        estado:
          type: string
          example: ACTIVO
        roles:
          example:
            - d5de12df-3cc3-5a58-a742-be24030482d8
          type: array
          items:
            type: string
        ciudadaniaDigital:
          type: boolean
          example: false
      required:
        - usuario
        - estado
        - roles
        - ciudadaniaDigital
    ActualizarContrasenaDto:
      type: object
      properties:
        contrasenaActual:
          type: string
          example: contrasena
        contrasenaNueva:
          type: string
          example: contrasena-nueva
      required:
        - contrasenaActual
        - contrasenaNueva
    ActualizarUsuarioRolDto:
      type: object
      properties:
        persona:
          allOf:
            - $ref: '#/components/schemas/PersonaDto'
        correoElectronico:
          type: object
          example: asdfg123@gmail.com
        roles:
          example:
            - '3'
          type: array
          items:
            type: string
      required:
        - persona
        - correoElectronico
        - roles
...